restructure project directories
 - restructure project in to a package

[Finishes package-structure]
