+++ flaskblog/__init__.py
@@ -1,0 +1,9 @@
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+
+app = Flask(__name__)
+app.config['SECRET_KEY'] = '9ddd8eb9de0def83119d5aca5504d771'
+app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///site.db'
+db = SQLAlchemy(app)
+
+from flaskblog import routes
