+++ requirements.txt
@@ -1,8 +1,10 @@
 Click==7.0
 Flask==1.1.1
+Flask-SQLAlchemy==2.4.1
 Flask-WTF==0.14.2
 itsdangerous==1.1.0
 Jinja2==2.10.3
 MarkupSafe==1.1.1
+SQLAlchemy==1.3.12
 Werkzeug==0.16.0
 WTForms==2.2.1
